requirements.txt
json
google-generativeai
pydantic
langgraph
langchain-core
langchain-google-genai

